# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Management System API",
        "description": "Etkinlik Yönetim Sistemi REST API Dokümantasyonu - Türkiye'nin önde gelen etkinlik yönetim platformu",
        "contact": {
            "name": "Event Management Support Team",
            "url": "https://eventmanagement.com/support",
            "email": "admin@eventmanagement.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Yerel Geliştirme Sunucusu"
        },
        {
            "url": "https://api.eventmanagement.com",
            "description": "Production API Server"
        }
    ],
    "paths": {
        "/api/v1/organizations": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Organizasyonları listele",
                "description": "Mevcut organizasyonları sayfalama, filtreleme ve arama ile birlikte listeler",
                "operationId": "5812c950de77aa5c38c80b05be6f9b52",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Arama terimi (isim, açıklama, email)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Durum filtresi",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "user_role",
                        "in": "query",
                        "description": "Kullanıcı rolü filtresi",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "editor",
                                "viewer"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sıralama alanı",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "name",
                            "enum": [
                                "name",
                                "created_at",
                                "events_count",
                                "participants_count",
                                "sponsors_count"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Sıralama yönü",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Sayfa başına öğe sayısı (max: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrganizationSummary"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Yeni organizasyon oluştur",
                "description": "Yeni bir organizasyon oluşturur ve mevcut kullanıcıyı admin olarak ekler",
                "operationId": "4676987fd0a00de6d8fa6acf6f54a21c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Organizasyon adı",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Akademik Kongreler Derneği"
                                    },
                                    "description": {
                                        "description": "Organizasyon açıklaması",
                                        "type": "string",
                                        "maxLength": 5000,
                                        "nullable": true
                                    },
                                    "contact_email": {
                                        "description": "İletişim email",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "contact_phone": {
                                        "description": "İletişim telefonu",
                                        "type": "string",
                                        "maxLength": 20,
                                        "nullable": true
                                    },
                                    "website_url": {
                                        "description": "Website URL",
                                        "type": "string",
                                        "format": "url",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "logo": {
                                        "description": "Logo dosyası (jpeg,png,jpg,gif - max: 2MB)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "description": "Aktif durumu",
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Organizasyon başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Organizasyon başarıyla oluşturuldu."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrganizationDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasyon hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/participants/search": {
            "get": {
                "tags": [
                    "Participants"
                ],
                "summary": "Katılımcı ara (autocomplete)",
                "description": "Katılımcı seçimi için arama yapar (max 20 sonuç)",
                "operationId": "95f2a15cdbd07a74972ea8ad9c9bde12",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Arama terimi (min: 2 karakter)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 100,
                            "minLength": 2
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Rol filtresi",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "speaker",
                                "moderator"
                            ]
                        }
                    },
                    {
                        "name": "organization_id",
                        "in": "query",
                        "description": "Organizasyon filtresi",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ParticipantSearchResult"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasyon hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/participants/import": {
            "post": {
                "tags": [
                    "Participants"
                ],
                "summary": "Katılımcı içe aktar",
                "description": "CSV/Excel dosyasından katılımcı içe aktarır",
                "operationId": "41541487fa74e21efd75e951f2dc6fec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "CSV/Excel dosyası (max: 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "organization_id": {
                                        "description": "Organizasyon ID",
                                        "type": "integer"
                                    },
                                    "mapping": {
                                        "description": "Alan eşleştirmeleri",
                                        "properties": {
                                            "first_name": {
                                                "description": "Ad alanı",
                                                "type": "string"
                                            },
                                            "last_name": {
                                                "description": "Soyad alanı",
                                                "type": "string"
                                            },
                                            "email": {
                                                "description": "Email alanı",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "update_existing": {
                                        "description": "Mevcut kayıtları güncelle",
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "501": {
                        "description": "Henüz implement edilmedi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Import functionality will be implemented in a dedicated ImportController."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrganizationDetail": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/OrganizationSummary"
                    }
                ]
            },
...
            "ValidationError": {
                "title": "Validation Error",
                "description": "Doğrulama hatası yanıtı",
                "properties": {
                    "success": {
                        "description": "İşlem başarı durumu",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Hata mesajı",
                        "type": "string",
                        "example": "Doğrulama hatası"
                    },
                    "errors": {
                        "description": "Alan bazlı hatalar",
                        "properties": {
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Name alanı zorunludur."
                                }
                            },
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Geçerli bir e-posta adresi giriniz."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Error": {
                "title": "Error Response",
                "description": "Hata yanıtı",
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "description": "İşlem başarı durumu",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Hata mesajı",
                        "type": "string",
                        "example": "Bir hata oluştu"
                    },
                    "error_code": {
                        "description": "Hata kodu",
                        "type": "string",
                        "example": "ORG_001",
                        "nullable": true
                    },
                    "details": {
                        "description": "Hata detayları",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Laravel Sanctum Bearer Token Authentication"
            },
            "api_key": {
                "type": "apiKey",
                "description": "API Key Authentication",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}